+++ src/main/java/com/alpha/redux/well/enchants/LifestealLore.java
@@ -8,7 +8,7 @@ public class LifestealLore extends PitEnchant{
 
     @Override
     public void run(ReduxDamageEvent event, int level) {
-        int multiplier = 0;
+        double multiplier = 0;
 
         if (level > 2) {
             multiplier += (level*4) + 1;
